+++ src/main/java/paramsForRequests/ParamsForRequests.java
@@ -329,90 +329,5 @@ public class ParamsForRequests extends MainParamsForRequest {
             e.printStackTrace();
         }
     }
-
-    @Step
-    public String getLZTicketForSale(int ticketCount, int parochkaCount, String bTV){
-        String ticketPart1 = "{\"tickets\":[{\"t\":";
-        String ticketPart2 = ",\"f\":0,\"e\":[{\"c\":1,\"n\":";
-        String ticketPart3 = "},{\"c\":2,\"n\":";
-        String ticketPart4 = "}]}]}";
-        String richAndPopular;
-        boolean isStateCheck = bTV.toLowerCase().equals("check");
-        boolean isStateUncheck = bTV.toLowerCase().equals("uncheck");
-        if (isStateCheck) {
-            richAndPopular = "1";
-        }
-        else if (isStateUncheck) {
-            richAndPopular = "0";
-        }
-        else{
-            richAndPopular = "0";
-        }
-        return  ticketPart1 + ticketCount + ticketPart2 + parochkaCount
-                + ticketPart3 + richAndPopular + ticketPart4;
-    }
-
-    //"{\"tickets\":[{\"t\":1,\"f\":0,\"e\":[{\"c\":1,\"n\":";
-    //        String ticketPart3 = "},{\"c\":2,\"n\":0}]}]}";
-
-    @Step
-    private String calculateOneBetSum(int ticketCount, int parochkaCount, String bTV){
-        int richAndPopular;
-        boolean isStateCheck = bTV.toLowerCase().equals("check");
-        boolean isStateUncheck = bTV.toLowerCase().equals("uncheck");
-        if (isStateCheck) {
-            richAndPopular = 2;
-        }
-        else if (isStateUncheck) {
-            richAndPopular = 0;
-        }
-        else{
-            richAndPopular = 0;
-        }
-        int mainTicketSum = 20;
-        int parochkaSum = (ticketCount * parochkaCount * 5) / ticketCount;
-        int sumOfOneTicket = mainTicketSum + parochkaSum + richAndPopular;
-        return String.valueOf(sumOfOneTicket);
-    }
-
-//    @Step
-//    public String calculateCheckSum(int ticketCount, int parochkaCount, int bTV){
-//        int oneBetSum = Integer.parseInt(calculateOneBetSum(ticketCount, parochkaCount, bTV));
-//        int result = oneBetSum * ticketCount;
-//        return result + "00";
-//    }
-
-    @Step
-    public String calculateLZCheckSumHrn(int ticketCount, int parochkaCount, String bTV){
-        int richAndPopular = 0;
-        boolean isStateCheck = bTV.toLowerCase().equals("check");
-        boolean isStateUncheck = bTV.toLowerCase().equals("uncheck");
-        if (isStateCheck) {
-            richAndPopular = 2;
-        }
-        else if (isStateUncheck) {
-            richAndPopular = 0;
-        }
-        int mainTicketSum = ticketCount * 20;
-        int parochkaSum = ticketCount * parochkaCount * 5;
-        float betSum = mainTicketSum + parochkaSum + (richAndPopular * ticketCount);
-        return String.format(Locale.ROOT,"%.2f", betSum);
-    }
-
-    @Step
-    public String calculateLZCheckSumMonets(int ticketCount, int parochkaCount, String bTV){
-        int richAndPopular = 0;
-        boolean isStateCheck = bTV.toLowerCase().equals("check");
-        boolean isStateUncheck = bTV.toLowerCase().equals("uncheck");
-        if (isStateCheck) {
-            richAndPopular = 2;
-        }
-        else if (isStateUncheck) {
-            richAndPopular = 0;
-        }
-        int mainTicketSum = ticketCount * 20;
-        int parochkaSum = ticketCount * parochkaCount * 5;
-        return String.valueOf(mainTicketSum + parochkaSum + (richAndPopular * ticketCount)) + "00";
-    }
 }
